remove circular dependencies
